backend/requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
cryptography==36.0.2
Flask==2.0.3
Werkzeug==2.2.2
gunicorn

# Testing dependencies
pytest==7.4.3
pytest-flask==1.3.0
pytest-mock==3.12.0
pytest-cov==4.1.0